plugins/towncrier/tests/fixtures/python-mid/pyproject.toml
[project]
name = "fixture-python-mid"
version = "0.0.0"
description = "Towncrier python-mid calibration fixture (mixed-quality)."
dependencies = [
  "structlog>=23.0",
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.20",
  "opentelemetry-sdk>=1.20",
  "fastapi>=0.110",
]
